fix: unknown error
missing head branch

+++ frontend/lib/core/widgets/overview/summary_card.dart
@@ -4,15 +4,9 @@ class SummaryCard extends StatelessWidget {
   final bool type;
   final String title;
   final double amount;
-  final int items;
+  final double items;
   final double avgPerMonth;
-  const SummaryCard(
-      {super.key,
-      required this.type,
-      required this.title,
-      required this.amount,
-      required this.items,
-      required this.avgPerMonth});
+  const SummaryCard({super.key, required this.type, required this.title, required this.amount, required this.items, required this.avgPerMonth});
 
   @override
   Widget build(BuildContext context) {
@@ -34,8 +28,7 @@ class SummaryCard extends StatelessWidget {
               const SizedBox(height: 5),
               Text(
                 amount.toStringAsFixed(2),
-                style: TextStyle(
-                    fontSize: 13, color: type ? Colors.green : Colors.red),
+                style: TextStyle(fontSize: 13, color: type ? Colors.green : Colors.red),
               ),
               Row(
                 children: [
@@ -46,8 +39,7 @@ class SummaryCard extends StatelessWidget {
                   const Spacer(),
                   Text(
                     items.toStringAsFixed(0),
-                    style: TextStyle(
-                        fontSize: 9, color: type ? Colors.green : Colors.red),
+                    style: TextStyle(fontSize: 9, color: type ? Colors.green : Colors.red),
                   ),
                 ],
               ),
@@ -60,8 +52,7 @@ class SummaryCard extends StatelessWidget {
                   const Spacer(),
                   Text(
                     avgPerMonth.toStringAsFixed(2),
-                    style: TextStyle(
-                        fontSize: 9, color: type ? Colors.green : Colors.red),
+                    style: TextStyle(fontSize: 9, color: type ? Colors.green : Colors.red),
                   ),
                 ],
               ),
